// src/client/types.gen.ts
export type LoginDto = {
  /**
   * 用户名
   */
  username: string
  /**
   * 密码
   */
  password: string
}